# Set the excluded files list once when getting lint files