Fixes #5329 

### Issue

The SPM plugin is much slower than the regular SwiftLint.

### Analysis

The plugin specifies not just an input folder but a list of all the project files. `getFiles()` resolves the excluded files for each of the specified files.

### Solution

No need to resolve the excluded files for each input file. Let's set them before iterating lint files.