@SimplyDanny 2023-11-05
Interesting. I would expect the compiler to extract code that is not influenced by variables being looped over.

Could you verify that performance is again on par with command line linting?