@vladvlasov256 2023-11-05
@SimplyDanny, sure, here is my test:

**Prerequisites**:

Not empty `ruby_env` is excluded:
```yaml
# .swiftlint.yml
excluded:
  - ruby_env
  - "**/.build"
```

I installed some gems there via `bundle config set --local path ruby_env  && bundle install`.

A test file, `FigletTool.swift`:
```swift
import Figlet

@main
struct FigletTool {
  static func main() {
    Figlet.say("Hello, Swift!")
  }
}
```

**0.53.0**:

```bash
time  swiftlint lint --quiet FigletTool.swift  
swiftlint lint --quiet   0.15s user 0.14s system 101% cpu 0.279 total

time swiftlint lint --quiet \                
FigletTool.swift FigletTool.swift FigletTool.swift FigletTool.swift FigletTool.swift FigletTool.swift FigletTool.swift FigletTool.swift \
FigletTool.swift FigletTool.swift FigletTool.swift FigletTool.swift FigletTool.swift FigletTool.swift FigletTool.swift FigletTool.swift \
FigletTool.swift FigletTool.swift FigletTool.swift FigletTool.swift FigletTool.swift FigletTool.swift FigletTool.swift FigletTool.swift \
FigletTool.swift FigletTool.swift FigletTool.swift FigletTool.swift FigletTool.swift FigletTool.swift FigletTool.swift FigletTool.swift \
FigletTool.swift FigletTool.swift FigletTool.swift FigletTool.swift FigletTool.swift FigletTool.swift FigletTool.swift FigletTool.swift \
FigletTool.swift FigletTool.swift FigletTool.swift FigletTool.swift FigletTool.swift FigletTool.swift FigletTool.swift FigletTool.swift \
FigletTool.swift FigletTool.swift FigletTool.swift FigletTool.swift FigletTool.swift FigletTool.swift FigletTool.swift FigletTool.swift \
FigletTool.swift FigletTool.swift FigletTool.swift FigletTool.swift FigletTool.swift FigletTool.swift FigletTool.swift FigletTool.swift \
FigletTool.swift FigletTool.swift FigletTool.swift FigletTool.swift FigletTool.swift FigletTool.swift FigletTool.swift FigletTool.swift \
FigletTool.swift FigletTool.swift FigletTool.swift FigletTool.swift FigletTool.swift FigletTool.swift FigletTool.swift FigletTool.swift \
FigletTool.swift FigletTool.swift FigletTool.swift FigletTool.swift FigletTool.swift FigletTool.swift FigletTool.swift FigletTool.swift \
FigletTool.swift FigletTool.swift FigletTool.swift FigletTool.swift FigletTool.swift FigletTool.swift FigletTool.swift FigletTool.swift \
FigletTool.swift FigletTool.swift FigletTool.swift FigletTool.swift FigletTool.swift FigletTool.swift FigletTool.swift FigletTool.swift \
FigletTool.swift FigletTool.swift FigletTool.swift FigletTool.swift FigletTool.swift FigletTool.swift FigletTool.swift FigletTool.swift \
FigletTool.swift FigletTool.swift FigletTool.swift FigletTool.swift FigletTool.swift FigletTool.swift FigletTool.swift FigletTool.swift

swiftlint lint --quiet             6.34s user 14.47s system 104% cpu 19.872 total
```

**After the fix**:
```bash
time swiftlint lint --quiet FigletTool.swift 
swiftlint lint --quiet   0.15s user 0.14s system 101% cpu 0.278 total


time swiftlint lint --quiet \                
FigletTool.swift FigletTool.swift FigletTool.swift FigletTool.swift FigletTool.swift FigletTool.swift FigletTool.swift FigletTool.swift \
FigletTool.swift FigletTool.swift FigletTool.swift FigletTool.swift FigletTool.swift FigletTool.swift FigletTool.swift FigletTool.swift \
FigletTool.swift FigletTool.swift FigletTool.swift FigletTool.swift FigletTool.swift FigletTool.swift FigletTool.swift FigletTool.swift \
FigletTool.swift FigletTool.swift FigletTool.swift FigletTool.swift FigletTool.swift FigletTool.swift FigletTool.swift FigletTool.swift \
FigletTool.swift FigletTool.swift FigletTool.swift FigletTool.swift FigletTool.swift FigletTool.swift FigletTool.swift FigletTool.swift \
FigletTool.swift FigletTool.swift FigletTool.swift FigletTool.swift FigletTool.swift FigletTool.swift FigletTool.swift FigletTool.swift \
FigletTool.swift FigletTool.swift FigletTool.swift FigletTool.swift FigletTool.swift FigletTool.swift FigletTool.swift FigletTool.swift \
FigletTool.swift FigletTool.swift FigletTool.swift FigletTool.swift FigletTool.swift FigletTool.swift FigletTool.swift FigletTool.swift \
FigletTool.swift FigletTool.swift FigletTool.swift FigletTool.swift FigletTool.swift FigletTool.swift FigletTool.swift FigletTool.swift \
FigletTool.swift FigletTool.swift FigletTool.swift FigletTool.swift FigletTool.swift FigletTool.swift FigletTool.swift FigletTool.swift \
FigletTool.swift FigletTool.swift FigletTool.swift FigletTool.swift FigletTool.swift FigletTool.swift FigletTool.swift FigletTool.swift \
FigletTool.swift FigletTool.swift FigletTool.swift FigletTool.swift FigletTool.swift FigletTool.swift FigletTool.swift FigletTool.swift \
FigletTool.swift FigletTool.swift FigletTool.swift FigletTool.swift FigletTool.swift FigletTool.swift FigletTool.swift FigletTool.swift \
FigletTool.swift FigletTool.swift FigletTool.swift FigletTool.swift FigletTool.swift FigletTool.swift FigletTool.swift FigletTool.swift \
FigletTool.swift FigletTool.swift FigletTool.swift FigletTool.swift FigletTool.swift FigletTool.swift FigletTool.swift FigletTool.swift

swiftlint lint --quiet             0.16s user 0.15s system 108% cpu 0.286 total
```

Without the fix, it's approximately linear time for "1 file vs 120 files". And the constant time with the fix.

So there should be an excluded directory to traverse to reproduce slow performance. If I remove gems the test takes 1.6s. But even in this case, the fixed version needs only 0.127s for execution.