@SimplyDanny 2023-11-06
A performance test would be cool. I assume the number of files in `ruby_env` is quite high though. The test would have to generate them.

Even though I have never used them, Xcode provides a few methods for performance testing. They seem to be helpful here. Do you want to try adding an integration test?